Remove now unused footer buttons
